# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the nursery management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a private object upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL and normalized object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "401":
          description: Unauthorized
        "400":
          description: Invalid upload metadata
        "404":
          description: Application not found
        "409":
          description: Accepted applications cannot be edited
        "415":
          description: Unsupported document type
...
components:
...
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [applicationId, name, size, contentType]
      properties:
        applicationId: { type: number, minimum: 1, multipleOf: 1 }
        name: { type: string, minLength: 1 }
        size: { type: number, minimum: 1, maximum: 10485760, multipleOf: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl: { type: string }
        objectPath:
          type: string
          pattern: "^/objects/uploads/[0-9a-fA-F]{8}-(?:[0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}$"
...